requirements.txt
cairocffi==1.7.1
CairoSVG==2.7.1
cffi==1.16.0
chess==1.10.0
cssselect2==0.7.0
defusedxml==0.7.1
pillow==10.4.0
pycparser==2.22
python-chess==1.999
tinycss2==1.3.0
webencodings==0.5.1
